Cleaning up & Warning fixes
