--- src/main/java/com/spring/demo/models/Employees.java
@@ -1,50 +1,0 @@
-package com.spring.demo.models;
-
-import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
-@Table(name = "employee")
-public class Employees extends Humans {
-
-    @Id
-    @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
-    @GeneratedValue(generator = "uuid2")
-    @Column(length = 64)
-    private String id;
-
-    private String field;
-
-    @OneToMany(fetch = FetchType.LAZY)
-    private List<Shipment> shipment;
-
-    public String getField() {
-        return field;
-    }
-
-    public void setField(String field) {
-        this.field = field;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public List<Shipment> getShipment() {
-        return shipment;
-    }
-
-    public void setShipment(List<Shipment> shipment) {
-        this.shipment = shipment;
-    }
-}
