--- src/main/java/com/spring/demo/models/Tags.java
@@ -1,47 +1,0 @@
-package com.spring.demo.models;
-
-
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-@Table(name = "tag")
-public class Tags {
-
-    @Id
-    @GenericGenerator(name = "uuid2", strategy = "org.hibernate.id.UUIDGenerator")
-    @GeneratedValue(generator = "uuid2")
-    @Column(length = 64)
-    private String id;
-
-    private String tag;
-
-    @ManyToMany(mappedBy = "tags")
-    private List<Product> products;
-
-    public String getId() {
-        return id;
-    }
-
-    public String getTag() {
-        return tag;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public void setTag(String tag) {
-        this.tag = tag;
-    }
-
-    public List<Product> getProducts() {
-        return products;
-    }
-
-    public void setProducts(List<Product> products) {
-        this.products = products;
-    }
-}
