+++ src/main/java/com/spring/demo/repositories/DeliveryRepository.java
@@ -1,7 +1,7 @@
 package com.spring.demo.repositories;
 
-import com.spring.demo.models.Deliveries;
+import com.spring.demo.models.Delivery;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeliveryRepository extends CrudRepository<Deliveries, String> {
+public interface DeliveryRepository extends CrudRepository<Delivery, String> {
 }
