+++ src/main/java/com/spring/demo/repositories/EmployeeRepository.java
@@ -1,7 +1,7 @@
 package com.spring.demo.repositories;
 
-import com.spring.demo.models.Employees;
+import com.spring.demo.models.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepository extends CrudRepository<Employees, String> {
+public interface EmployeeRepository extends CrudRepository<Employee, String> {
 }
