+++ src/main/java/com/spring/demo/repositories/TagRepository.java
@@ -1,13 +1,13 @@
 package com.spring.demo.repositories;
 
-import com.spring.demo.models.Tags;
+import com.spring.demo.models.Tag;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface TagRepository extends CrudRepository<Tags, String> {
+public interface TagRepository extends CrudRepository<Tag, String> {
 
-    @Query(value = "select * from Tag tag where tag.name like %?1")
-    Optional<Tags> getTagByName(String tagName);
+    @Query(value = "select tag from Tag tag where tag.tag like %?1")
+    Optional<Tag> getTagByName(String tagName);
 }
