+++ src/main/java/com/spring/demo/repositories/UserRepository.java
@@ -1,7 +1,7 @@
 package com.spring.demo.repositories;
 
-import com.spring.demo.models.Users;
+import com.spring.demo.models.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<Users, String> {
+public interface UserRepository extends CrudRepository<User, String> {
 }
