+++ src/main/resources/application.properties
@@ -5,3 +5,4 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
+spring.jpa.open-in-view=false
